+++ payment-service/src/main/java/com/hedvig/paymentservice/services/swish/PayoutPayload.kt
@@ -1,0 +1,16 @@
+package com.hedvig.paymentservice.services.swish
+
+data class PayoutPayload(
+    val payerAlias: String,
+    val payoutInstructionUUID: String,
+    val payerPaymentReference: String,
+    val payeeAlias: String,
+    val payeeSSN: String,
+    val amount: String,
+    val currency: String,
+    val message: String,
+    val instructionDate: String,
+    val signingCertificateSerialNumber: String
+) {
+    val payoutType: String = "PAYOUT"
+}
