+++ payment-service/src/main/java/com/hedvig/paymentservice/services/swish/PayoutRequest.kt
@@ -1,0 +1,7 @@
+package com.hedvig.paymentservice.services.swish
+
+data class PayoutRequest(
+    val payload: PayoutPayload,
+    val signature: String,
+    val callbackUrl: String
+)
