+++ payment-service/src/main/java/com/hedvig/paymentservice/services/swish/SwishService.kt
@@ -3,8 +3,6 @@ package com.hedvig.paymentservice.services.swish
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.hedvig.paymentservice.services.swish.client.SwishClient
 import feign.FeignException
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import java.time.LocalDateTime
 import java.util.UUID
@@ -26,11 +24,13 @@ class SwishService(
         instructionDate: LocalDateTime,
     ): StartPayoutResponse {
         val payload = PayoutPayload(
+            payerAlias = properties.payerAlias,
             payoutInstructionUUID = transactionId.toString().replace("-", "").toUpperCase(),
             payerPaymentReference = memberId,
             payeeAlias = payeeAlias,
             payeeSSN = payeeSSN,
             amount = String.format("%.2f", amount),
+            currency = amount.currency.currencyCode,
             message = message,
             instructionDate = instructionDate.toString(),
             signingCertificateSerialNumber = properties.signingCertificateSerialNumber
@@ -45,25 +45,5 @@ class SwishService(
             StartPayoutResponse.Failed(e.message, (e as FeignException?)?.status())
         }
     }
-
-    data class PayoutRequest(
-        val payload: PayoutPayload,
-        val signature: String,
-        val callbackUrl: String
-    )
-
-    data class PayoutPayload(
-        val payoutInstructionUUID: String,
-        val payerPaymentReference: String,
-        val payeeAlias: String,
-        val payeeSSN: String,
-        val amount: String,
-        val message: String,
-        val instructionDate: String,
-        val signingCertificateSerialNumber: String
-    ) {
-        val payerAlias: String = "1235261086"
-        val currency: String = "SEK"
-        val payoutType: String = "PAYOUT"
-    }
 }
+
